requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
pyodbc==5.0.1
pymssql==2.2.11
psycopg2-binary==2.9.9

# OpenAI
openai==1.3.7

# Microsoft Teams Bot
botbuilder-core==4.17.0
botbuilder-schema==4.17.0
aiohttp==3.9.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2

# Logging & Monitoring
loguru==0.7.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
